core/tx_pool: reviewed tx pool use cases
